validaciones, permisos

+++ app/Controllers/Categoria.php
@@ -12,7 +12,18 @@ class Categoria extends BaseController{
         $data['paginador']=$paginador;
         $titulo = "Categorias";
         $data['titulo'] = $titulo;
-        return view('productos/categorias', $data);
+        
+        //validacion rol permisos
+        $session = session();
+        if ($session->has('id_rol')) {
+            $rol_usuario = $session->get('id_rol');
+        }
+        if (in_array($rol_usuario, [1,2])) {
+            return view('productos/categorias', $data);
+        }else{
+            $data['titulo'] = 'Error 404';
+            return view('errors/html/error_404', $data);
+        }
     }
 
     public function guardar(){
@@ -34,8 +45,19 @@ class Categoria extends BaseController{
             'descripcion'=>$this->request->getVar('descripcion'),
             'estado'=>$estado
         ];
-        $categoria->insert($datos);
-        return $this->response->redirect(site_url('categorias'));
+        
+        //validacion rol permisos
+        $session = session();
+        if ($session->has('id_rol')) {
+            $rol_usuario = $session->get('id_rol');
+        }
+        if (in_array($rol_usuario, [1,2])) {
+            $categoria->insert($datos);
+            return $this->response->redirect(site_url('categorias'));
+        }else{
+            $data['titulo'] = 'Error 404';
+            return view('errors/html/error_404', $data);
+        }
     }
 
     public function actualizar(){
@@ -59,30 +81,44 @@ class Categoria extends BaseController{
             //return $this->response->redirect(site_url('/listar'));
             return redirect()->back()->withInput();
         }
-
-        $categoria->update($id,$data);
-
-        return $this->response->redirect(site_url('categorias'));
+        
+        //validacion rol permisos
+        $session = session();
+        if ($session->has('id_rol')) {
+            $rol_usuario = $session->get('id_rol');
+        }
+        if (in_array($rol_usuario, [1,2])) {
+            $categoria->update($id,$data);
+            return $this->response->redirect(site_url('categorias'));
+        }else{
+            $data['titulo'] = 'Error 404';
+            return view('errors/html/error_404', $data);
+        }
     }
     
     public function eliminar($id = null) {
-
         
-        $producto = new Productos();
-        //validar si hay categorias con productos relacionados
-        $producto_cat = $producto->where('id_categoria', $id)->findAll();
-        if(!empty($producto_cat)){
-            $session = session();
-            $session->setFlashData('mensaje','No se puede eliminar la categoria ya que tiene productos relacionados.');
-            return $this->response->redirect(site_url('categorias'));
-        } else {
-            $categoria = new Categorias();
-            $categoria->where('id', $id)->delete($id);
-            return $this->response->redirect(site_url('categorias'));
+        //validacion rol permisos
+        $session = session();
+        if ($session->has('id_rol')) {
+            $rol_usuario = $session->get('id_rol');
+        }
+        if (in_array($rol_usuario, [1,2])) {
+            $producto = new Productos();
+            //validar si hay categorias con productos relacionados
+            $producto_cat = $producto->where('id_categoria', $id)->findAll();
+            if(!empty($producto_cat)){
+                $session = session();
+                $session->setFlashData('mensaje','No se puede eliminar la categoria ya que tiene productos relacionados.');
+                return $this->response->redirect(site_url('categorias'));
+            } else {
+                $categoria = new Categorias();
+                $categoria->where('id', $id)->delete($id);
+                return $this->response->redirect(site_url('categorias'));
+            }
+        }else{
+            $data['titulo'] = 'Error 404';
+            return view('errors/html/error_404', $data);
         }
     }
-
-    
-
-
 }
