+++ app/Controllers/Producto.php
@@ -85,7 +85,18 @@ class Producto extends BaseController{
                                     ->paginate(10);
         $paginador = $producto->pager;
         $data['paginador']=$paginador;
-        return view('productos/index', $data);
+        
+        //validacion rol permisos
+        $session = session();
+        if ($session->has('id_rol')) {
+            $rol_usuario = $session->get('id_rol');
+        }
+        if (in_array($rol_usuario, [1,2,3])) {
+            return view('productos/index', $data);
+        }else{
+            $data['titulo'] = 'Error 404';
+            return view('errors/html/error_404', $data);
+        }
     }
     
     public function crear(){
@@ -93,7 +104,18 @@ class Producto extends BaseController{
         $data['categorias'] = $categoria->orderBy('id','ASC')->findAll();
         $titulo = "Productos";
         $data['titulo'] = $titulo;
-        return view('productos/crear', $data);
+        
+        //validacion rol permisos
+        $session = session();
+        if ($session->has('id_rol')) {
+            $rol_usuario = $session->get('id_rol');
+        }
+        if (in_array($rol_usuario, [1,2])) {
+            return view('productos/crear', $data);
+        }else{
+            $data['titulo'] = 'Error 404';
+            return view('errors/html/error_404', $data);
+        }
     }
 
     public function guardar(){
@@ -172,11 +194,19 @@ class Producto extends BaseController{
 
         $producto = new Productos();
         $datos = $producto->where('id', $id)->first();
-        $producto->where('id', $id)->delete($id);
-        return $this->response->redirect(site_url('productos'));
-
         
-
+        //validacion rol permisos
+        $session = session();
+        if ($session->has('id_rol')) {
+            $rol_usuario = $session->get('id_rol');
+        }
+        if (in_array($rol_usuario, [1,2])) {
+            $producto->where('id', $id)->delete($id);
+            return $this->response->redirect(site_url('productos'));
+        }else{
+            $data['titulo'] = 'Error 404';
+            return view('errors/html/error_404', $data);
+        }
     }
 
     public function editar($id=null){
@@ -186,7 +216,18 @@ class Producto extends BaseController{
         $data['categorias'] = $categoria->orderBy('id','ASC')->findAll();
         $titulo = "Productos";
         $data['titulo'] = $titulo;
-        return view('productos/editarProducto', $data);
+        
+        //validacion rol permisos
+        $session = session();
+        if ($session->has('id_rol')) {
+            $rol_usuario = $session->get('id_rol');
+        }
+        if (in_array($rol_usuario, [1,2])) {
+            return view('productos/editarProducto', $data);
+        }else{
+            $data['titulo'] = 'Error 404';
+            return view('errors/html/error_404', $data);
+        }
     }
 
     public function actualizarProducto(){
@@ -236,112 +277,122 @@ class Producto extends BaseController{
     }
     
     public function generarExcel($nombre, $codigo, $estado, $inventariable, $iva, $descuento, $categoriaFiltro){
-        $producto = new Productos();
-        // Aplica los filtros solo si se han seleccionado valores
-        if ($nombre != 'none' || $codigo != 'none' || $estado != 'none' || $inventariable != 'none' || $iva != 'none' || $descuento != 'none' || $categoriaFiltro != 'none') {
-            $producto->orderBy('id', 'ASC');
-
-            if ($nombre != 'none') {
-                $producto->like('nombre', $nombre);
-            }
-
-            if ($codigo != 'none') {
-                $producto->like('codigo', $codigo);
-            }
-
-            if ($estado != 'none') {
-                $producto->where('productos.estado', $estado);
-            }
-
-            if ($inventariable != 'none') {
-                $producto->where('es_inventariable', $inventariable);
-            }
-
-            if ($iva != 'none') {
-                $producto->where('porcentaje_iva', $iva);
-            }
-
-            
-            if ($descuento != 'none') {
-                if ($descuento != '0'){
-                    $producto->where('descuento !=', '0');
-                }
-                else{
-                    $producto->where('descuento', $descuento);
-                }
-            }
-
-            if ($categoriaFiltro != 'none') {
-                $producto->where('id_categoria', $categoriaFiltro);
-            }
-
-
-            $data['productos'] = $producto->select('productos. *, categoria_producto.descripcion as categoria')
-                                    ->join('categoria_producto', 'categoria_producto.id = productos.id_categoria', 'left')
-                                    ->findAll();
-        } else {
-            $data['productos'] = $producto->select('productos. *, categoria_producto.descripcion as categoria')
-                                    ->join('categoria_producto', 'categoria_producto.id = productos.id_categoria', 'left')
-                                    ->findAll();
+        
+        //validacion rol permisos
+        $session = session();
+        if ($session->has('id_rol')) {
+            $rol_usuario = $session->get('id_rol');
         }
-
-        // Crear un nuevo objeto Spreadsheet
-        $spreadsheet = new Spreadsheet();
-
-        // Obtener la hoja activa
-        $sheet = $spreadsheet->getActiveSheet();
-
-        $sheet->getColumnDimension('C')->setWidth(35);
-        $sheet->getColumnDimension('D')->setWidth(20);  
-        $sheet->getColumnDimension('E')->setWidth(15);  
-        $sheet->getColumnDimension('H')->setWidth(15);  
-        $sheet->getColumnDimension('I')->setWidth(15);  
-        $sheet->getColumnDimension('J')->setWidth(20);  
-        // Agregar encabezados
-        $sheet->setCellValue('A1', 'ID');
-        $sheet->setCellValue('B1', 'Codigo');
-        $sheet->setCellValue('C1', 'Nombre');
-        $sheet->setCellValue('D1', 'Categoría');
-        $sheet->setCellValue('E1', 'Inventariable');
-        $sheet->setCellValue('F1', 'Stock');
-        $sheet->setCellValue('G1', 'IVA');
-        $sheet->setCellValue('H1', 'Precio Venta');
-        $sheet->setCellValue('I1', 'Precio Compra');
-        $sheet->setCellValue('J1', 'Porcentaje Descuento');
-        $sheet->setCellValue('K1', 'Estado');
-        // ... Agregar más columnas según tus necesidades
-
-        // Agregar datos de las personas al archivo Excel
-        $row = 2;
-        foreach ($data['productos'] as $producto) {
-            $sheet->setCellValue('A' . $row, $producto['id']);
-            $sheet->setCellValue('B' . $row, $producto['codigo']);
-            $sheet->setCellValue('C' . $row, $producto['nombre']);
-            $sheet->setCellValue('D' . $row, $producto['categoria']);
-            $sheet->setCellValue('E' . $row, $producto['es_inventariable'] == '1' ? 'Si' : 'No' );
-            $sheet->setCellValue('F' . $row, $producto['stock']);
-            $sheet->setCellValue('G' . $row, $producto['porcentaje_iva'].'%');
-            $sheet->setCellValue('H' . $row, $producto['precio_venta']);
-            $sheet->setCellValue('I' . $row, $producto['precio_compra']);
-            $sheet->setCellValue('J' . $row, $producto['descuento']);
-            $sheet->setCellValue('K' . $row, $producto['estado'] == '1' ? 'Activo' : 'Inactivo');
+        if (in_array($rol_usuario, [1,2])) {
+        
+            $producto = new Productos();
+            // Aplica los filtros solo si se han seleccionado valores
+            if ($nombre != 'none' || $codigo != 'none' || $estado != 'none' || $inventariable != 'none' || $iva != 'none' || $descuento != 'none' || $categoriaFiltro != 'none') {
+                $producto->orderBy('id', 'ASC');
+    
+                if ($nombre != 'none') {
+                    $producto->like('nombre', $nombre);
+                }
+    
+                if ($codigo != 'none') {
+                    $producto->like('codigo', $codigo);
+                }
+    
+                if ($estado != 'none') {
+                    $producto->where('productos.estado', $estado);
+                }
+    
+                if ($inventariable != 'none') {
+                    $producto->where('es_inventariable', $inventariable);
+                }
+    
+                if ($iva != 'none') {
+                    $producto->where('porcentaje_iva', $iva);
+                }
+    
+                
+                if ($descuento != 'none') {
+                    if ($descuento != '0'){
+                        $producto->where('descuento !=', '0');
+                    }
+                    else{
+                        $producto->where('descuento', $descuento);
+                    }
+                }
+    
+                if ($categoriaFiltro != 'none') {
+                    $producto->where('id_categoria', $categoriaFiltro);
+                }
+    
+    
+                $data['productos'] = $producto->select('productos. *, categoria_producto.descripcion as categoria')
+                                        ->join('categoria_producto', 'categoria_producto.id = productos.id_categoria', 'left')
+                                        ->findAll();
+            } else {
+                $data['productos'] = $producto->select('productos. *, categoria_producto.descripcion as categoria')
+                                        ->join('categoria_producto', 'categoria_producto.id = productos.id_categoria', 'left')
+                                        ->findAll();
+            }
+    
+            // Crear un nuevo objeto Spreadsheet
+            $spreadsheet = new Spreadsheet();
+    
+            // Obtener la hoja activa
+            $sheet = $spreadsheet->getActiveSheet();
+    
+            $sheet->getColumnDimension('C')->setWidth(35);
+            $sheet->getColumnDimension('D')->setWidth(20);  
+            $sheet->getColumnDimension('E')->setWidth(15);  
+            $sheet->getColumnDimension('H')->setWidth(15);  
+            $sheet->getColumnDimension('I')->setWidth(15);  
+            $sheet->getColumnDimension('J')->setWidth(20);  
+            // Agregar encabezados
+            $sheet->setCellValue('A1', 'ID');
+            $sheet->setCellValue('B1', 'Codigo');
+            $sheet->setCellValue('C1', 'Nombre');
+            $sheet->setCellValue('D1', 'Categoría');
+            $sheet->setCellValue('E1', 'Inventariable');
+            $sheet->setCellValue('F1', 'Stock');
+            $sheet->setCellValue('G1', 'IVA');
+            $sheet->setCellValue('H1', 'Precio Venta');
+            $sheet->setCellValue('I1', 'Precio Compra');
+            $sheet->setCellValue('J1', 'Porcentaje Descuento');
+            $sheet->setCellValue('K1', 'Estado');
             // ... Agregar más columnas según tus necesidades
-
-            $row++;
+    
+            // Agregar datos de las personas al archivo Excel
+            $row = 2;
+            foreach ($data['productos'] as $producto) {
+                $sheet->setCellValue('A' . $row, $producto['id']);
+                $sheet->setCellValue('B' . $row, $producto['codigo']);
+                $sheet->setCellValue('C' . $row, $producto['nombre']);
+                $sheet->setCellValue('D' . $row, $producto['categoria']);
+                $sheet->setCellValue('E' . $row, $producto['es_inventariable'] == '1' ? 'Si' : 'No' );
+                $sheet->setCellValue('F' . $row, $producto['stock']);
+                $sheet->setCellValue('G' . $row, $producto['porcentaje_iva'].'%');
+                $sheet->setCellValue('H' . $row, $producto['precio_venta']);
+                $sheet->setCellValue('I' . $row, $producto['precio_compra']);
+                $sheet->setCellValue('J' . $row, $producto['descuento']);
+                $sheet->setCellValue('K' . $row, $producto['estado'] == '1' ? 'Activo' : 'Inactivo');
+                // ... Agregar más columnas según tus necesidades
+    
+                $row++;
+            }
+    
+            // Guardar el archivo Excel
+            $writer = new Xlsx($spreadsheet);
+            $filename = 'export_productos.xlsx';
+            $writer->save($filename);
+    
+            // Descargar el archivo Excel
+            header('Content-Type: application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
+            header('Content-Disposition: attachment;filename="' . $filename . '"');
+            header('Cache-Control: max-age=0');
+            $writer->save('php://output');
+            exit;
+        }else{
+            $data['titulo'] = 'Error 404';
+            return view('errors/html/error_404', $data);
         }
-
-        // Guardar el archivo Excel
-        $writer = new Xlsx($spreadsheet);
-        $filename = 'export_productos.xlsx';
-        $writer->save($filename);
-
-        // Descargar el archivo Excel
-        header('Content-Type: application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
-        header('Content-Disposition: attachment;filename="' . $filename . '"');
-        header('Cache-Control: max-age=0');
-        $writer->save('php://output');
-        exit;
     }
-
-
 }
